Renderer: Implement Metal texture and pipeline

+++ Engine/Include/Renderer/OpenGL/OpenGLFramebuffer.hpp
@@ -13,7 +13,7 @@ public:
     virtual void Unbind() override;
     virtual void Resize(uint32_t width, uint32_t height) override;
 
-    virtual uint32_t GetColorAttachmentRendererID() const override { return m_ColorAttachment; }
+    virtual uintptr_t GetColorAttachmentRendererID() const override { return m_ColorAttachment; }
     virtual const FramebufferSpecification &GetSpecification() const override { return m_Specification; }
 
 private:
